back-end/requirements.txt
attrs==22.2.0
beautifulsoup4==4.11.1
branca==0.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.2
folium==0.14.0
geographiclib==2.0
geopy==2.3.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
packaging==23.0
plotly==5.13.1
pluggy==1.0.0
psycopg2-binary==2.9.5
py==1.11.0
pytest==7.1.3
python-dotenv==0.20.0
requests==2.28.2
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
tenacity==8.2.2
tomli==2.0.1
urllib3==1.26.14
Werkzeug==2.2.2
